% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3pOperations.R
\name{preProcess_cropWS}
\alias{preProcess_cropWS}
\alias{x3p_cropWS}
\title{Crop rows/columns of missing values around an x3p}
\usage{
x3p_cropWS(x3p, croppingThresh = 1, ...)
}
\arguments{
\item{x3p}{an x3p object containing a surface matrix}

\item{croppingThresh}{minimum number of non-NA pixels that need to be in a
row/column for it to not be cropped out of the surface matrix}

\item{crop}{}
}
\value{
a surface matrix with outer rows/columns removed depending on
  croppingThresh
}
\description{
Crop rows/columns of missing values around an x3p
}
\examples{
\dontrun{
raw_x3p <- x3ptools::read_x3p("path/to/file.x3p") \%>\%
  x3ptools::sample_x3p(m = 2)

raw_x3p$surface.matrix <- raw_x3p$surface.matrix \%>\%
  cmcR::preProcess_ransacLevel() \%>\%
  cmcR::preProcess_levelBF() \%>\%
  cmcR::preProcess_cropWS(croppingThresh = 2)
}

}
