% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3pOperations.R
\name{x3p_cropWS}
\alias{x3p_cropWS}
\title{Crop rows/columns of missing values around an x3p}
\usage{
x3p_cropWS(x3p, croppingThresh = 1, ...)
}
\arguments{
\item{x3p}{an x3p object containing a surface matrix}

\item{croppingThresh}{minimum number of non-NA pixels that need to be in a
row/column for it to not be cropped out of the surface matrix}

\item{crop}{}
}
\value{
a surface matrix with outer rows/columns removed depending on
  croppingThresh
}
\description{
Crop rows/columns of missing values around an x3p
}
\examples{
data("K013sA1")

x3pCropped <- x3p_cropWS(K013sA1,croppingThresh = 5)

x3pPlot(K013sA1,x3pCropped)

}
