% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3pOperations.R
\name{x3p_delete}
\alias{x3p_delete}
\title{Replace elements of a surface matrix based on a mask}
\usage{
x3p_delete(x3p, mask_vals, replacement = NA)
}
\arguments{
\item{x3p}{an x3p object}

\item{mask_vals}{a hexidecimal color value that corresponds to indices in a
mask whose indices are to be replaced in the associated surface
matrix}

\item{replacement}{value used replace indices in the surface matrix}
}
\description{
Replace elements of a surface matrix based on a mask
}
