% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3pOperations.R
\name{x3p_elemAverage}
\alias{x3p_elemAverage}
\title{Calculate the element-wise average between two x3ps}
\usage{
x3p_elemAverage(x3p1, x3p2)
}
\arguments{
\item{x3p1}{an x3p object}

\item{x3p2}{another x3p object}
}
\value{
An x3p object containing the element-wise average between the surface
  values of x3p1 and x3p1
}
\description{
Calculates the element-wise average between the surface values
  of two x3ps (of the same dimension)
}
