% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3pOperations.R
\name{x3p_sd}
\alias{x3p_sd}
\title{Calculate the standard deviation of x3p surface values}
\usage{
x3p_sd(..., na.rm = TRUE)
}
\arguments{
\item{...}{one or more x3p objects}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
The standard deviation of the input surface matrices
}
\description{
Calculates the standard deviation of all surface values of one
  or more x3p objects
}
\examples{
data("K013sA1","K013sA2")

# calculates the sd for a single x3p's surface values
x3p_sd(K013sA1)

# calculates the sd for the joint surface values for two x3ps
x3p_sd(K013sA1,K013sA2)

}
