% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{x3p_to_dataFrame}
\alias{x3p_to_dataFrame}
\title{Convert an x3p object to a data frame}
\usage{
x3p_to_dataFrame(x3p, preserveResolution = FALSE)
}
\arguments{
\item{x3p}{an x3p object}

\item{preserveResolution}{a boolean dictating whether the scan resolution is
preserved in the returned data frame. If FALSE, then the x,y data frame
columns will be integer-valued. Otherwise, the difference between
consecutive x,y values will be equal to the scan resolution.}
}
\description{
Convert an x3p object to a data frame
}
