% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisonPlot.R
\name{x3p_comparisonPlot}
\alias{x3p_comparisonPlot}
\title{Create plots of filtered element-wise average and differences}
\usage{
x3p_comparisonPlot(
  x3p1,
  x3p2,
  plotNames = c("x3p1", "x3p2", "Element-wise Average", "x3p1 diff.", "x3p2 diff."),
  type = "faceted",
  thresholdMultiplier = 1,
  unit = "Norm."
)
}
\description{
Create plots of filtered element-wise average and differences
}
